@@ -10,33 +10,33 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@mdx-js/loader": "2.3.0",
-    "@mdx-js/react": "2.3.0",
-    "@next/font": "13.1.1",
-    "@next/mdx": "13.4.1",
-    "@types/bootstrap": "5.2.1",
-    "@types/platform": "1.3.4",
-    "@types/react-bootstrap": "0.32.30",
-    "@types/uuid": "9.0.0",
-    "bootstrap": "5.2.0",
-    "bootstrap-icons": "1.9.1",
+    "@mdx-js/loader": "3.0.1",
+    "@mdx-js/react": "3.0.1",
+    "@next/font": "14.2.4",
+    "@next/mdx": "14.2.4",
+    "@types/bootstrap": "5.2.10",
+    "@types/platform": "1.3.6",
+    "@types/react-bootstrap": "0.32.36",
+    "@types/uuid": "10.0.0",
+    "bootstrap": "5.3.3",
+    "bootstrap-icons": "1.11.3",
     "he": "1.2.0",
-    "next": "13.1.1",
-    "next-i18next": "12.1.0",
-    "next-language-detector": "1.0.2",
+    "next": "14.2.4",
+    "next-i18next": "15.3.0",
+    "next-language-detector": "1.1.0",
     "platform": "1.3.6",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "typewriter-effect": "2.19.0",
-    "uuid": "9.0.0"
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "typewriter-effect": "2.21.0",
+    "uuid": "10.0.0"
   },
   "devDependencies": {
-    "@types/node": "18.6.3",
-    "@types/react": "18.0.15",
-    "@types/react-dom": "18.0.6",
+    "@types/node": "20.14.8",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
     "eslint": "8.21.0",
-    "eslint-config-next": "^13.1.1",
+    "eslint-config-next": "13.1.1",
     "http-server": "14.1.1",
-    "typescript": "4.7.4"
+    "typescript": "5.5.2"
   }
 }
